requirements.txt
cocotb==1.9.2
matplotlib==3.9.2
numpy==2.1.3
PyYAML==6.0.2
